Add Combatant quantities (2nd time).
The first time I did this was on another branch (either "quantities" or
"dictionaries," I can't figure out which. I then tried rewinding...
...and anyway, here's the result. Moving on. Doing quantities as dictionaries
to the Encounter object was the wrong move. This seems more hack-ish at first
glance, but it's better. (Combatant objects now have a _quantity inst var that
keeps track of how many to incarnate for each combat.)